requirements.txt
requests>=2.32.3
python-dotenv>=1.0.1
PyJWT>=2.9.0
google-genai>=1.5.0
rich>=13.9.4
json-repair>=0.30.0
